+++ banzai/dbs.py
@@ -454,13 +454,17 @@ def get_master_calibration_image(image, calibration_type, master_selection_crite
 
 
 def get_individual_calibration_images(instrument, calibration_type, min_date: datetime.datetime,
-                                      max_date: datetime.datetime, use_masters=False, db_address=_DEFAULT_DB):
+                                      max_date: datetime.datetime, use_masters=False, include_bad_frames=False,
+                                      db_address=_DEFAULT_DB):
 
     calibration_criteria = CalibrationImage.instrument_id == instrument.id
     calibration_criteria &= CalibrationImage.type == calibration_type.upper()
     calibration_criteria &= CalibrationImage.is_master.is_(use_masters)
     calibration_criteria &= CalibrationImage.dateobs >= min_date
     calibration_criteria &= CalibrationImage.dateobs <= max_date
+
+    if not include_bad_frames:
+        calibration_criteria &= CalibrationImage.is_bad == False
 
     with get_session(db_address=db_address) as db_session:
         images = db_session.query(CalibrationImage).filter(calibration_criteria).all()