+++ CHANGES.md
@@ -2,6 +2,7 @@
 -------------------
 - Fixed a bug in stacking master calibrations from the command line.
 - Fixed logging format to conform to LCO standard format and to not be overridden by celery
+- Fixed a bug that would use calibration frames to stack even if they were marked as bad
 
 0.23.0 (2019-05-08)
 -------------------